requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
httpx==0.25.0
openai==1.30.5
aiogram==3.0.0
python-dotenv==1.0.0
loguru==0.7.0
# For OpenAI API v1.x, httpx>=0.27.0 and openai>=1.3.0 are required
matplotlib==3.8.2
aiohttp>=3.8.5  # Updated to match aiogram's requirement
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
python-telegram-bot-pagination==0.0.3
chromedriver-autoinstaller==0.6.4
tenacity==8.2.0  # For retries
prometheus-client==0.17.0  # For metrics
sentry-sdk==1.39.0  # For error tracking
psutil==5.9.8  # For system metrics